requirements.txt
aioredis==1.3.1
anyio==3.6.2
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
bidict==0.22.1
certifi==2022.12.7
cffi==1.15.1
channels==3.0.1
channels-redis==4.1.0
charset-normalizer==3.1.0
constantly==15.1.0
cryptography==40.0.2
daphne==3.0.2
defusedxml==0.7.1
dj-static==0.0.6
Django==4.2
django-allauth==0.54.0
django-cors-headers==3.14.0
django-environ==0.10.0
django-filter==23.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
environ==1.0
h11==0.14.0
h2==4.1.0
hiredis==2.2.3
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
Markdown==3.4.3
methoddispatch==3.0.2
msgpack==1.0.5
oauthlib==3.2.2
Pillow==9.5.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.1.1
python-engineio==4.4.1
python3-openid==3.2.0
pytz==2023.3
redis==4.5.5
requests==2.31.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
static-ranges==0.2.0
static3==0.7.0
Twisted==22.10.0
twisted-iocpsupport==1.0.3
txaio==23.1.1
typing_extensions==4.6.2
tzdata==2023.3
urllib3==1.26.15
whitenoise==6.4.0
zope.interface==6.0
